refactor(account): lift Account data + achievements toast into hooks (R2)
Extract the four header reads (username/profile/counts/gamification) + derived
display strings into useAccountData, and the newly-unlocked achievements toast
side effect into useAchievementsToast. Account.tsx 204 -> 126 lines, now
presentational. Behaviour-preserving (Account suite 20/20).
